requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart
llama-cpp-python
faiss-cpu
sentence-transformers